Test cases

+++ include/body.hpp
@@ -8,10 +8,10 @@ class Body {
     Vector2 pos;
     Vector2 vel;
     Vector2 acc;
-    double mass;
+    float mass;
 
     Body(Vector2 arg_pos, Vector2 arg_vel, double arg_mass);
 
-    void update(double dt);
+    void update(float dt);
 };
 }  // namespace sim
